+++ package.json
@@ -20,7 +20,8 @@
     "react-dom": "18.2.0",
     "react-helmet-async": "1.3.0",
     "react-redux": "8.1.3",
-    "react-router-dom": "6.16.0"
+    "react-router-dom": "6.16.0",
+    "react-toastify": "^10.0.5"
   },
   "devDependencies": {
     "@jedmao/redux-mock-store": "3.0.5",
