+++ src/components/bookmark/bookmark.tsx
@@ -1,4 +1,12 @@
 import cn from 'classnames';
+import { useEffect, useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+
+import { ShortOfferType } from '../../types';
+import { AppRoutesEnum } from '../../consts';
+import { getFavorites } from '../../store/offers/selectors';
+import { setIsFavoriteAction } from '../../store/api-actions';
+import { useAppDispatch, useAppSelector, useIsAuthorized } from '../../hooks';
 
 const SIZES = {
   BIG: { width: '31', height: '33' },
@@ -6,30 +14,41 @@ const SIZES = {
 };
 
 type Props = {
-  isActive: boolean;
   className: string;
-  onClick: () => void;
   size?: keyof typeof SIZES;
+  offer: Pick<ShortOfferType, 'id' | 'isFavorite'>;
 };
 
-export function Bookmark({
-  onClick,
-  isActive,
-  className,
-  size = 'SMALL',
-}: Props) {
+export function Bookmark({ offer, className, size = 'SMALL' }: Props) {
+  const navigate = useNavigate();
+  const dispatch = useAppDispatch();
+  const isAuthorized = useIsAuthorized();
+  const favorites = useAppSelector(getFavorites);
+  const [isFavorite, setIsFavorite] = useState(offer.isFavorite);
+  const onClick = () => {
+    if (isAuthorized) {
+      dispatch(setIsFavoriteAction({ offerId: offer.id, status: !isFavorite }));
+    } else {
+      navigate(AppRoutesEnum.LOGIN);
+    }
+  };
+
+  useEffect(() => {
+    setIsFavorite(favorites.some((item) => item.id === offer.id));
+  }, [favorites, offer.id]);
+
   return (
     <button
       type="button"
       onClick={onClick}
       className={cn(`${className}__bookmark-button button`, {
-        [`${className}__bookmark-button--active`]: isActive,
+        [`${className}__bookmark-button--active`]: isFavorite,
       })}
     >
       <svg
-        className={`${className}__bookmark-icon`}
         width={SIZES[size].width}
         height={SIZES[size].height}
+        className={`${className}__bookmark-icon`}
       >
         <use xlinkHref="#icon-bookmark"></use>
       </svg>
